Commonize interface generators
