+++ command/command_freebsd.go
@@ -27,6 +27,7 @@ func metricsGenerators(conf *config.Config) []metrics.Generator {
 		&metricsFreebsd.CPUUsageGenerator{},
 		&metrics.FilesystemGenerator{IgnoreRegexp: conf.Filesystems.Ignore.Regexp, UseMountpoint: conf.Filesystems.UseMountpoint},
 		&metricsFreebsd.MemoryGenerator{},
+		&metrics.InterfaceGenerator{Interval: metricsInterval},
 	}
 
 	return generators
