--- metrics/darwin/interface.go
@@ -1,72 +1,0 @@
-// +build darwin
-
-package darwin
-
-import (
-	"time"
-
-	"github.com/mackerelio/go-osstat/network"
-	"github.com/mackerelio/golib/logging"
-	"github.com/mackerelio/mackerel-agent/metrics"
-	"github.com/mackerelio/mackerel-agent/util"
-)
-
-/*
-collect network interface I/O
-
-`interface.{interface}.{metric}.delta`: The increased amount of network I/O per minute retrieved from the result of netstat -bni
-
-interface = "en0", "en1" and so on...
-*/
-
-// InterfaceGenerator XXX
-type InterfaceGenerator struct {
-	Interval time.Duration
-}
-
-// metrics for posting to Mackerel
-
-var interfaceLogger = logging.GetLogger("metrics.interface")
-
-// Generate XXX
-func (g *InterfaceGenerator) Generate() (metrics.Values, error) {
-	prevValues, err := g.collectInterfacesValues()
-	if err != nil {
-		return nil, err
-	}
-
-	time.Sleep(g.Interval)
-
-	currValues, err := g.collectInterfacesValues()
-	if err != nil {
-		return nil, err
-	}
-
-	ret := make(map[string]float64)
-	for name, value := range prevValues {
-		currValue, ok := currValues[name]
-		if ok {
-			ret[name+".delta"] = float64(currValue-value) / g.Interval.Seconds()
-		}
-	}
-
-	return metrics.Values(ret), nil
-}
-
-func (g *InterfaceGenerator) collectInterfacesValues() (map[string]uint64, error) {
-	networks, err := network.Get()
-	if err != nil {
-		interfaceLogger.Errorf("failed to get network statistics: %s", err)
-		return nil, err
-	}
-	if len(networks) == 0 {
-		return nil, nil
-	}
-	results := make(map[string]uint64, len(networks)*2)
-	for _, network := range networks {
-		name := util.SanitizeMetricKey(network.Name)
-		results["interface."+name+".rxBytes"] = network.RxBytes
-		results["interface."+name+".txBytes"] = network.TxBytes
-	}
-	return results, nil
-}
