+++ metrics/interface.go
@@ -1,13 +1,12 @@
-// +build linux
+// +build !windows
 
-package linux
+package metrics
 
 import (
 	"time"
 
 	"github.com/mackerelio/go-osstat/network"
 	"github.com/mackerelio/golib/logging"
-	"github.com/mackerelio/mackerel-agent/metrics"
 	"github.com/mackerelio/mackerel-agent/util"
 )
 
@@ -16,7 +15,7 @@ collect network interface I/O
 
 `interface.{interface}.{metric}.delta`: The increased amount of network I/O per minute retrieved from /proc/net/dev
 
-interface = "eth0", "eth1" and so on...
+interface = "eth0", "eth1" and so on... ("en0" on darwin)
 */
 
 // InterfaceGenerator generates interface metric values
@@ -27,7 +26,7 @@ type InterfaceGenerator struct {
 var interfaceLogger = logging.GetLogger("metrics.interface")
 
 // Generate interface metric values
-func (g *InterfaceGenerator) Generate() (metrics.Values, error) {
+func (g *InterfaceGenerator) Generate() (Values, error) {
 	prevValues, err := g.collectInterfacesValues()
 	if err != nil {
 		return nil, err
@@ -47,7 +46,7 @@ func (g *InterfaceGenerator) Generate() (metrics.Values, error) {
 		}
 	}
 
-	return metrics.Values(ret), nil
+	return Values(ret), nil
 }
 
 func (g *InterfaceGenerator) collectInterfacesValues() (map[string]uint64, error) {
